Refactor Admin Panel: Update component imports and remove unused files
- Updated import paths for ContentCreator and Profile components to reflect new directory structure.
- Deleted obsolete ContentCreater, Profile, and logout components to streamline the codebase.
- Adjusted Adminpanel index to incorporate changes in component structure.
