--- Client/src/pages/Adminpanel/Adminpages/Contents/logout.tsx
@@ -1,11 +1,0 @@
-import React from 'react'
-
-function logout() {
-  return (
-    <div>
-      logout
-    </div>
-  )
-}
-
-export default logout
